Groundwork for new treaties
This commit lays the groundwork for adding treaties. It modifies the test files and documentation to recognize "treaty" as a valid cite_type. It also updates the note on for the Statutes and Large to indicate that some of those citations may be to treaties if they predate 1949. I opted not to modify the sources which could be statutes or treaties (statutes at large and Lexis) because most citations from those sources are not to treaties.