Added treaties
This commit adds the Indigo Book T6 treaty sources to reporters-db. Treaty sources are stored in either the laws.json or reporters.json file depending on which style they more closely resemble. Regardless of where they are stored, they all have cite_type "treaty." I omitted two sources (Statutes at Large and LEXIS) which are not exclusive to treaties and were already present in reporters-db. Also, the regex and example for Treaty Series (T.S.) begin with a white space. This is to prevent false positives where T.S. is part of the name of another reporter, but it is an inelegant solution. I tried various lookbehinds and negative assertions, but because they are at the beginning of the string they cause problems. All examples for all treaties are from Court Listener except for the example for Pan-Am. Treaty Series, which is from a law review article.