Creating Universal Section Marker and Applying it to U.S. Code
This commit creates a section_marker which can be called in any regex and which will match singular and plural section markers in a variety of forms (e.g. Sec., sections, and §§). It also uses that section_marker in the U.S. Code regex. If that works well, we can use it in the other regexes too.

@flooie I have to sign off for the day but I think this addresses all of your comments. Feel free to change the regex if you decide to do so. Note that I used double escape chars because otherwise the tests failed on my machine. I also made each of the groups you suggested into non-capture groups in case anybody uses selects groups by numbers (esp. now that the section_marker will obscure how many groups it might create for a given citation). Finally, I made "ec" mandatory rather than optional so that the letter S or s alone does not match (which it did for your draft regex).